requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.20.2
idna==3.6
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.2.0.post0
lightning-utilities==0.10.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nlpaug==1.1.11
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacremoses==0.1.1
safetensors==0.4.1
SciencePlots==2.1.1
scipy==1.12.0
seaborn==0.13.1
sentencepiece==0.1.99
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
timm==0.9.12
tokenizers==0.15.2
torch==2.1.2
torchmetrics==1.3.1
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.3
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
wandb==0.16.3
wcwidth==0.2.13
yarl==1.9.4
